Move CCommandArgs into its own header.
